Kernel: Split the SysFS core files into smaller components
