+++ Kernel/Bus/PCI/Controller/MemoryBackedHostBridge.h
@@ -10,6 +10,7 @@
 #include <AK/Vector.h>
 #include <Kernel/Bus/PCI/Controller/HostController.h>
 #include <Kernel/Locking/Spinlock.h>
+#include <Kernel/Memory/Region.h>
 
 namespace Kernel::PCI {
