+++ Kernel/FileSystem/SysFS/Subsystems/Firmware/BIOS/Directory.h
@@ -8,7 +8,6 @@
 
 #include <AK/Types.h>
 #include <AK/Vector.h>
-#include <Kernel/FileSystem/SysFS.h>
 #include <Kernel/FileSystem/SysFS/Subsystems/Firmware/Directory.h>
 #include <Kernel/Library/LockRefPtr.h>
 #include <Kernel/PhysicalAddress.h>
